Module rweb::filters

Built-in Filters

This module mostly serves as documentation to group together the list of built-in filters. Most of these are available at more convenient paths.

Modules

addr
Socket Address filters.

any
A filter that matches any route.

body
Body filters

cookie
Cookie Filters

cors
CORS Filters

ext
Request Extensions

fs
File System Filters

header
Header Filters

log
Logger Filters

method
HTTP Method filters.

multipart
Multipart body filters

path
Path Filters

query
Query Filters

reply
Reply Filters

sse
Server-Sent Events (SSE)

ws
Websockets Filters

Structs

BoxedFilter
A type representing a boxed Filter trait object.